+++ src/Spreads.Collections/Collections/ImmutableIntMap32U.fs
@@ -1252,17 +1252,12 @@ namespace Spreads.Collections
             member this.GetEnumerator() = this.GetCursor() :> IAsyncEnumerator<KVP<uint32, 'T>>
             member this.GetCursor() = this.GetCursor()
             member this.IsEmpty = this.IsEmpty
-
             member this.IsIndexed with get() = false
-
-            member this.First
-                with get() = this.First
-
-            member this.Last
-                with get() = this.Last
-
-            member this.Item
-                with get (k) : 'T = this.Item(k)
+            member this.IsMutable = false
+            member this.First with get() = this.First
+            member this.Last with get() = this.Last
+            member this.Item with get (k) : 'T = this.Item(k)
+            member this.GetAt(idx:int) : 'T = this.Skip(Math.Max(0, idx-1)).First().Value
             member this.Keys with get() = IntMap32UTree.keys tree :> IEnumerable<uint32>
             member this.Values with get() = IntMap32UTree.values tree :> IEnumerable<'T>