+++ src/Spreads.Collections/Collections/IndexedMap.fs
@@ -1,6 +1,7 @@
 ﻿namespace Spreads.Collections
 
 open System
+open System.Linq
 open System.Diagnostics
 open System.Collections
 open System.Collections.Generic
@@ -49,6 +50,8 @@ type IndexedMap<'K,'V> // when 'K:equality
 
   [<NonSerializedAttribute>]
   let mutable isSynchronized : bool = false
+  [<NonSerializedAttribute>]
+  let mutable isMutable : bool = true
   [<NonSerializedAttribute>]
   let syncRoot = new Object()
   [<NonSerializedAttribute>]
@@ -128,7 +131,13 @@ type IndexedMap<'K,'V> // when 'K:equality
     this.size <- this.size + 1
     if cursorCounter >0 then updateEvent.Trigger(KVP(k,v))
     
-  member internal this.IsReadOnly with get() = false
+  member this.IsMutable 
+    with get() = isMutable
+    and set (value) = 
+      if isMutable then isMutable <- value
+      else 
+        if isMutable = value then () // NB same as not value
+        else invalidOp "Cannot make immutable map mutable, the setter only supports on-way change from mutable to immutable"
 
   member this.IsSynchronized 
     with get() =  isSynchronized
@@ -839,7 +848,7 @@ type IndexedMap<'K,'V> // when 'K:equality
 
   interface IDictionary<'K,'V> with
     member this.Count = this.Count
-    member this.IsReadOnly with get() = this.IsReadOnly
+    member this.IsReadOnly with get() = not this.IsMutable
     member this.Item
       with get key = this.Item(key)
       and set key value = this.[key] <- value
@@ -874,6 +883,7 @@ type IndexedMap<'K,'V> // when 'K:equality
     member this.GetCursor() = this.GetCursor()
     member this.IsEmpty = this.size = 0
     member this.IsIndexed with get() = false
+    member this.IsMutable with get() = this.IsMutable
     member this.First with get() = this.First
     member this.Last with get() = this.Last
     member this.TryFind(k:'K, direction:Lookup, [<Out>] result: byref<KeyValuePair<'K, 'V>>) = 
@@ -896,6 +906,7 @@ type IndexedMap<'K,'V> // when 'K:equality
         false
     member this.TryGetValue(k, [<Out>] value:byref<'V>) = this.TryGetValue(k, &value)
     member this.Item with get k = this.Item(k)
+    member this.GetAt(idx:int) = this.values.[idx]
     member this.Keys with get() = this.Keys :> IEnumerable<'K>
     member this.Values with get() = this.values :> IEnumerable<'V>
     member this.SyncRoot with get() = syncRoot